stage1: Fix coremods being loaded before Mixin is initialized
Those mods which use Mixin may assume that Mixin is initialized before their
coremod is loaded.
With the current stage1 this was however not always the case: If the
MixinTweaker was already queued because another mod used it, then we do not
initialized it but we nevertheless add mixin containers for our mods
immediately, which does cause Mixin to partially initialize.
We also initialized the coremod of Essential mods twice, once ourselves and once
by adding the mixin container (which runs Mixin's coremod injection code).

With this commit, we skip our coremod code for Essential mods which use Mixin
and we start initializing Mixin before add our containers.