stage2/lw: Attempt to re-launch when old Kotlin/Mixin is loaded
That is, if we cannot work around the fact that an old version being on the
classpath by pre-loading our version (e.g. because the classes are already
loaded, or excluded from transformations), then we will attempt to "re-launch"
the game by spinning up an inner, isolated class loader which has our jar at the
beginning of the classpath, and then starting another Launch inside of that.

This way we can always get first on the classpath as long as the re-launching is
successful and not foiled by some global state (e.g. classes with native methods
cannot be loaded twice, or mods being annoying by trying really hard to never be
loaded twice).