stage2: Fix mod class being loaded by system class loader in dev env
This initial check - which we've inherited from the original loader, meant to
ensure that the mod is only initialized once (which is trivially the case now
anyway because the stage1 loader is already a singleton) - causes the class to be
loaded if it is already on the system classpath (e.g. because the Essential
dependency was not declared as `compileOnly`) which then prevents us from
loading our up-to-date version (because the one already loaded on the system
class loader takes priority).

This commit removes the initial check, the associated `isInitialized` method,
and replaces the cursed, legacy launchwrapper isInClassPath implementation with
the much cleaner one we have been using for Fabric.
Note that this is not actually sufficient to fix the issue in a LaunchWrapper
environment because we currently have to load our tweaker on the system class
loader for it to not break (see `EssentialLoader.getModClassLoader`).