stage2/lw: Add "meta" mixins
This is a separate instance of the whole Mixin subsystem able to transform
anything which would normally be loaded on the system class loader (but with
relaunch is loaded in the RelaunchClassLoader). This allows us to effectively
mixin into everything from coremods to FMLTweaker, even Launch and (the inner,
normal) Mixin.