stage2/lw: Use empty class loader as parent for IsolatedClassLoader
This is because Java 8 unconditionally inherits package definitions from the
parent class loader which we do not actually want. E.g. if we re-launched
because the ASM package is too old, we do not want that package to still be
present with the incorrect implementation version (even though the classes are
different) inside the relaunch class loader.