stage2: Store checksum in jar manifest
So we do not need to recompute it on every boot. This becomes important once we
implement updates via diff because then the read checksum of the resulting file
will no longer match the checksum of the file if we had downloaded it whole.

Also (preparing) to store some extra metadata in the manifest while we are at
it, mostly for diagnostic purposes and for when we support updating more than
just Essential.