stage2/ml9: Determine runtime MC version instead of hard-coded one
ModLauncher doesn't provide the MC version early enough for our loader to get it
during its constructor call. Until now, we've only supported a single version
and could simply side-step the issue by hard-coding it.
With this commit, the ShimEssentialLoader is what gets constructed first and the
actual loader isn't constructed until a later point at which we can read the MC
version.

Note: The ShimEssentialLoader needs to be renamed back to EssentialLoader,
because that's the class which stage1 will instantiate. That's done in a
separate commit, to keep the correct history on the ActualEssentialLoader file.