new: Support for pinning stage2 and stage3
Allowing Essential to be used in an offline / strict no-auto-update environment.

Also lays the initial foundations in stage2 for third-party container mods that
can load and auto-update a third-party mod just like we currently have the
Essential container mod.
Note that there's still quite a bit of work required for this to function
though. In particular there's various places in the platform-specific parts of
stage2 that currently assume that they will only ever deal with Essential.