new: Prevent soft-lock when pinning is used and mod is broken
This commit completely changes the way click-to-update works such that the
loader can detect if the mod has not made any decision (e.g. because its update
modal broke) and can then on the next boot offer the update via its own GUI to
prevent the user from getting stuck with an old, broken version.

Furthermore it also enables users to downgrade the jar that's in use by
downgrading the container mod.
It does however only do this when the previous pinned version is at least as new
as the jar in use, as to prevent accidental downgrade when the modpack author
upgrade the container mod to a version that's not quite the most recent one.

The update prompt GUI itself is still pending design.