stage1: Skip update checks when relaunching
Otherwise it'll appear to the inner instance as if the mod is broken because an
update was pending but the mod didn't do anything with it; and as such it'll
immediately open the fallback modal.

Even worse, if the user accepts, we'll fail to write the update to disk on
Windows because the stage2 jar file is already loaded at this point.
This also applies to the full update (in case a new version releases in the slim
window between outer and inner stage1 checking calling the api).