stage2: Store jar file metadata next to jar instead of inside
This way the checksum of the downloaded file can be preserved.
(Diff-updates do not currently preserve the checksum, this is addressed in the
next commit.)

This does have the small downside that the jar could be replaced without the
meta file being updated. But if the user messes with internal files, that's on
them.