stage2: Make diff updates deterministic
Provided that the mod jar itself has no timestamps and is sorted, the result
should be byte-for-byte identical to a full upgrade.
Of note, this sorting is absolute across all entries of the jar. Gradle by
default will only sort each `from` of which the jar is built but keep the
different `from`s in the order in which they were declared (I assume, that's
what it looks like at least). Hence why our test build script manually sorts our
dummy stage3.
Loom 1.3 on the other hand does sort all entries at the end of its `remapJar`
task (afaik, untested because Essential has additional build steps afterwards).

This commit also verifies that the old jar does indeed have the expected hash
before preforming a diff update, otherwise falling back to downloading a full
jar.
This is because we used to modify the downloaded jar in non-deterministic ways,
and need to start from a known good point to get good results.
It also guards against applying patches onto a bad base (e.g. because the user
installed 1.20 Essential on 1.20.1).

This commit also verifies that the output jar of the diff update has the
expected hash, otherwise falling back to downloading a full jar.
This is because as of the previous commit, we no longer store the expected hash
in the jar, so there's no way for the Essential mod to recognize the jar file
based on its hash if its actual hash isn't correct.

This does unfortunately also put the loader a step further away from being
useful for loading third-party mods, since it places an additional restriction
on them (or will otherwise always re-download the full jar on each update).
Whether requiring third-parties distributed via Essential infra to have
reproducible jars is a feature or an issue is to be determined at a later time.

This does mean however, that such a reproducible version of Essential should be
published before this loader version, as otherwise all Essential users will have
to re-download the full jar until such a version is published.