build: Use same version for all platforms within a given stage
While all the platforms are technically independent, usually when we change
something, it affects all of them. And even in the cases where it doesn't, not
much harm is done by an updated that doesn't change anything.
And having matching versions across all platforms simplifies version management
and the release process.