stage2/ml9: Automatically determine layer of SelfRenamingJarMetadata
Newer Forge versions will put inner jars (such as Kotlin inside KFF) on
a layer which depends on the outer jar (KFF), which may be different
than what one would expect from just looking at the Kotlin jar.
As such we can no longer determine the layer ahead of time and will
instead assume that layers are built in order, and just assume whichever
layer is the most recent one.
Note that we cannot easily just check which layer our SecureJar is in
because it may be further wrapped by e.g. KFFMerger and as such may not
directly appear in any of the lists at all.