stage2/ml9: Fix crash with SinytraConnector
The "Sinytra Connector" mod uses a custom `SecureJar` implementation,
which prior to this commit resulted in us doing an out of bounds memory
read (returning what looks like a String, resulting in a class cast
exception) because we were using the same unsafe getter that we were
using for ML's `Jar` implementation.

This commit fixes the issue by using a separate getter for each
implementation class.
This change alone is sufficient to fix the issue because we'll then get
an explicit `NoSuchFieldException`, which we catch. However, to avoid
unnessecary log spam, this commit additionally special cases the dummy
class to silently return `null`. It also changes our `getVersion` method
to still try the fallback paths (which do work!) instead of returning
early on `null` since that is now a legitimate case.