mixin: Add
This commit imports Essential's mixin patches which improve backwards
compatibility of Mixin 0.8.x with mods expecting Mixin 0.7.x internals
and with Minecraft 1.8.9 which has libs older than what Mixin 0.8 needs.

This allows third-party mods to use Mixin 0.8 on 1.8.9 and 1.12.2 that
way it previously was only possible to do with full Essential.