stage2/lw: Add builtin support for MixinExtras
That is, we'll automatically initialize it for mods which bundle it.

This is so mods which previously used Essential's relocated MixinExtras,
for which initialization was handled by Essential, without Essential
without any further changes.
And since we're at it, we'll also provide the same for upstream
MixinExtras to make using it convenient too.