tests/lw: Fix logging
This plugin database file exists in multiple input jars, and if we
simply pick one of them then we'll be missing most of the plugins (e.g.
the one that loads xml configs), breaking logging.
If we exclude the file, log4j will fallback to a different path with
does discover enough plugins for it to function correctly.