build/mixin: Handle bridge types in shadow method/field descriptor
E.g. the `inject(CallbackBridge)` in our `CallbackInjectorCompat` mixin
uses our `CallbackBridge` type instead of Mixin's `Callback` type
because the latter is private.
We already remap `CallbackBridge` to `Callback` everywhere via
`ClassRemapper` after merging the mixin into the target, but we also
need to do that when comparing type descriptors of methods/fields in the
mixin and the target class, otherwise we won't be able to find the
target method/field.