mixin: Add support for MixinBooter
This commit adds stubs and custom functionality as required by
MixinBooter to not crash when booting.

Mods which rely on other features only available in MixinBooter's Mixin
fork or which rely on the Mixin version shipped with MixinBooter, may of
course still not function properly. Although we are not currently aware
of any such mod.